got big gpt training

+++ nlp/torch-transformer/dataset-script/generate.py
@@ -18,10 +18,13 @@ for pages in range(0, 1_000):
         folder = f"/mnt/blockstorage/text-dataset/{bucket}"
         os.makedirs(folder, exist_ok=True)
         text_file = f"{folder}/{text_id}.txt"
+        if os.path.isfile(text_file):
+            continue
+        response = requests.get(f"{HOST}/text/{text_id}", cookies={
+            "credentials": os.environ["auth_header"]
+        })
+        text = response.text
+        if response.status_code != 200 or len(text) == 0:
+            continue
         with open(text_file, "w") as file:
-            response = requests.get(f"{HOST}/text/{text_id}", cookies={
-                "credentials": os.environ["auth_header"]
-            })
-            if response.status_code != 200:
-                continue
-            file.write(response.text)
+            file.write(text)
